fix: use originating URL instead of `redirect_url` in auth
